Moved config.c from config directory to fc directory
