+++ src/main/blackbox/blackbox.c
@@ -37,6 +37,7 @@
 #include "drivers/compass.h"
 #include "drivers/system.h"
 
+#include "fc/config.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
 
@@ -48,7 +49,6 @@
 #include "sensors/compass.h"
 #include "sensors/sonar.h"
 
-#include "config/config.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
 #include "config/feature.h"
