+++ src/main/config/feature.c
@@ -21,7 +21,6 @@
 
 #include "platform.h"
 
-#include "config/config.h"
 #include "config/config_master.h"
 #include "config/feature.h"
