+++ src/main/flight/failsafe.c
@@ -24,16 +24,18 @@
 
 #include "common/axis.h"
 
-#include "rx/rx.h"
 #include "drivers/system.h"
-#include "io/beeper.h"
-#include "io/motors.h"
+
+#include "fc/config.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
 
-#include "config/config.h"
-
 #include "flight/failsafe.h"
+
+#include "io/beeper.h"
+#include "io/motors.h"
+
+#include "rx/rx.h"
 
 /*
  * Usage:
