+++ src/main/io/ledstrip.c
@@ -44,7 +44,9 @@
 #include "common/axis.h"
 #include "common/utils.h"
 
+#include "fc/config.h"
 #include "fc/rc_controls.h"
+#include "fc/runtime_config.h"
 
 #include "sensors/battery.h"
 #include "sensors/sensors.h"
@@ -63,19 +65,16 @@
 #include "io/osd.h"
 #include "io/vtx.h"
 
-#include "rx/rx.h"
-
 #include "flight/failsafe.h"
 #include "flight/mixer.h"
 #include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/navigation.h"
 
+#include "rx/rx.h"
+
 #include "telemetry/telemetry.h"
 
-#include "fc/runtime_config.h"
-
-#include "config/config.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
 #include "config/feature.h"
