+++ src/main/io/osd.c
@@ -38,12 +38,12 @@
 #include "io/flashfs.h"
 #include "io/osd.h"
 
+#include "fc/config.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
 
 #include "flight/pid.h"
 
-#include "config/config.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
 #include "config/feature.h"
