+++ src/main/io/serial_cli.c
@@ -56,6 +56,7 @@ uint8_t cliMode = 0;
 #include "drivers/sdcard.h"
 #include "drivers/buf_writer.h"
 
+#include "fc/config.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
 
@@ -94,7 +95,6 @@ uint8_t cliMode = 0;
 #include "telemetry/telemetry.h"
 #include "telemetry/frsky.h"
 
-#include "config/config.h"
 #include "config/config_eeprom.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
