+++ src/main/main.c
@@ -63,8 +63,9 @@
 #include "bus_bst.h"
 #endif
 
-#include "fc/fc_msp.h"
+#include "fc/config.h"
 #include "fc/fc_tasks.h"
+#include "fc/fc_msp.h"
 #include "fc/rc_controls.h"
 #include "fc/runtime_config.h"
 
@@ -108,7 +109,6 @@
 #include "flight/failsafe.h"
 #include "flight/navigation.h"
 
-#include "config/config.h"
 #include "config/config_eeprom.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
