+++ src/main/telemetry/frsky.c
@@ -31,6 +31,8 @@
 #include "common/maths.h"
 #include "common/axis.h"
 
+#include "config/feature.h"
+
 #include "drivers/system.h"
 #include "drivers/sensor.h"
 #include "drivers/accgyro.h"
@@ -38,6 +40,9 @@
 #include "drivers/timer.h"
 #include "drivers/serial.h"
 
+#include "fc/config.h"
+#include "fc/rc_controls.h"
+#include "fc/runtime_config.h"
 
 #include "sensors/sensors.h"
 #include "sensors/acceleration.h"
@@ -46,20 +51,14 @@
 #include "sensors/battery.h"
 
 #include "io/serial.h"
-#include "fc/rc_controls.h"
 #include "io/gps.h"
-
-#include "rx/rx.h"
 
 #include "flight/mixer.h"
 #include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/altitudehold.h"
 
-#include "fc/runtime_config.h"
-
-#include "config/config.h"
-#include "config/feature.h"
+#include "rx/rx.h"
 
 #include "telemetry/telemetry.h"
 #include "telemetry/frsky.h"
