+++ src/main/telemetry/ltm.c
@@ -49,6 +49,10 @@
 #include "drivers/serial.h"
 #include "drivers/pwm_rx.h"
 
+#include "fc/config.h"
+#include "fc/rc_controls.h"
+#include "fc/runtime_config.h"
+
 #include "sensors/sensors.h"
 #include "sensors/acceleration.h"
 #include "sensors/gyro.h"
@@ -57,7 +61,6 @@
 #include "sensors/battery.h"
 
 #include "io/serial.h"
-#include "fc/rc_controls.h"
 #include "io/gimbal.h"
 #include "io/gps.h"
 #include "io/ledstrip.h"
@@ -76,9 +79,6 @@
 
 #include "telemetry/telemetry.h"
 #include "telemetry/ltm.h"
-
-#include "config/config.h"
-#include "fc/runtime_config.h"
 
 
 #define TELEMETRY_LTM_INITIAL_PORT_MODE MODE_TX
